movies_service/requirements.txt
fastapi==0.111.0
elasticsearch[async]==8.13.2
redis==4.5.3
orjson==3.10.15
python-decouple==3.8
faker==19.2.0
tenacity==9.0.0
pydantic-settings==2.8.1
pydantic~=2.10.6
more-itertools~=10.6.0
elastic-transport~=8.17.0
pytest==6.1.2
pytest-asyncio==0.12.0
aiohttp==3.11.12
python-multipart>=0.0.7
ujson!=4.0.2,!=4.1.0,!=4.2.0,!=4.3.0,!=5.0.0,!=5.1.0,>=4.0.1
uvicorn[standard]>=0.12.0
starlette==0.37.2
httpx>=0.23.0
fastapi-cli>=0.0.2
jinja2>=2.11.2
typing-extensions>=4.8.0
email-validator>=2.0.0
async-timeout==4.0.3
python-dateutil>=2.4
python-dotenv>=0.21.0
iniconfig
toml
py>=1.8.2
attrs>=17.4.0
packaging
pluggy==0.12
multidict==4.5
chardet==3.0.4
SQLAlchemy==2.0.39
Werkzeug==3.1.3
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
flake8==7.1.2
pre-commit==4.1.0
alembic==1.15.2
limits==4.7.3